+++ src/main/java/repository/MotoboyRepository.java
@@ -18,7 +18,7 @@ public class MotoboyRepository {
         boolean sair = false;
 
         while (!sair) {
-            String[] options = { "Adicionar", "Editar", "Remover", "Consultar", "Voltar" };
+            String[] options = {"Adicionar", "Editar", "Remover", "Consultar", "Voltar"};
 
             int opcao = JOptionPane.showOptionDialog(null, "Escolha uma opção para motoboy: ",
                     "Menu Motoboy", JOptionPane.DEFAULT_OPTION, JOptionPane.PLAIN_MESSAGE, null, options, options[0]);
@@ -69,57 +69,73 @@ public class MotoboyRepository {
     }
 
     private static void editar(String id) {
-        if (motoboyExiste(id)) {
-            String novoNome = JOptionPane.showInputDialog("Informe o novo nome:");
-            String novoCpf = JOptionPane.showInputDialog("Informe o novo CPF:");
-            String novoTelefone = JOptionPane.showInputDialog("Informe o novo telefone:");
-            String novaPlaca = JOptionPane.showInputDialog("Informe a nova placa da moto:");
-            String disponivel = JOptionPane.showInputDialog("Informe a nova disponibilidade (S ou N):");
+        if (isValidObjectId(id)) {
+            if (motoboyExiste(id)) {
+                String novoNome = JOptionPane.showInputDialog("Informe o novo nome:");
+                String novoCpf = JOptionPane.showInputDialog("Informe o novo CPF:");
+                String novoTelefone = JOptionPane.showInputDialog("Informe o novo telefone:");
+                String novaPlaca = JOptionPane.showInputDialog("Informe a nova placa da moto:");
+                String disponivel = JOptionPane.showInputDialog("Informe a nova disponibilidade (S ou N):");
 
-            Document motoboyAtualizado = new Document()
-                    .append("nome", novoNome)
-                    .append("cpf", novoCpf)
-                    .append("telefone", novoTelefone)
-                    .append("placaMoto", novaPlaca)
-                    .append("disponivel", disponivel);
+                Document motoboyAtualizado = new Document()
+                        .append("nome", novoNome)
+                        .append("cpf", novoCpf)
+                        .append("telefone", novoTelefone)
+                        .append("placaMoto", novaPlaca)
+                        .append("disponivel", disponivel);
 
-            collection.replaceOne(getFiltro(id), motoboyAtualizado);
+                collection.replaceOne(getFiltro(id), motoboyAtualizado);
 
-            JOptionPane.showMessageDialog(null, "Motoboy editado com sucesso!");
+                JOptionPane.showMessageDialog(null, "Motoboy editado com sucesso!");
+            } else {
+                JOptionPane.showMessageDialog(null, "Motoboy não encontrado.");
+            }
         } else {
-            JOptionPane.showMessageDialog(null, "Motoboy não encontrado.");
+            JOptionPane.showMessageDialog(null, "ID inválido. Certifique-se de fornecer um Object ID válido.");
         }
     }
 
+
     private static void remover(String id) {
-        if (motoboyExiste(id)) {
-            collection.deleteOne(getFiltro(id));
-            JOptionPane.showMessageDialog(null, "Motoboy removido com sucesso!");
-        } else {
-            JOptionPane.showMessageDialog(null, "Motoboy não encontrado.");
+        if(isValidObjectId(id)) {
+            if (motoboyExiste(id)) {
+                collection.deleteOne(getFiltro(id));
+                JOptionPane.showMessageDialog(null, "Motoboy removido com sucesso!");
+            } else {
+                JOptionPane.showMessageDialog(null, "Motoboy não encontrado.");
+            }
+        } else {JOptionPane.showMessageDialog(null, "ID inválido. Certifique-se de fornecer um Object ID válido.");
         }
+
     }
 
     private static void consultar() {
         FindIterable<Document> documentos = collection.find();
 
         for (Document documento : documentos) {
-            String id = documento.getObjectId("_id").toString();
-            String nome = documento.getString("nome");
-            String cpf = documento.getString("cpf");
-            String telefone = documento.getString("telefone");
-            String placaMoto = documento.getString("placaMoto");
-            String disponivel = documento.getString("disponivel");
+            String idDocumento = documento.getObjectId("_id").toString();
 
-            System.out.println("Identificador: " + id);
-            System.out.println("Nome: " + nome);
-            System.out.println("CPF: " + cpf);
-            System.out.println("Telefone: " + telefone);
-            System.out.println("Placa da moto: " + placaMoto);
-            System.out.println("Disponível: " + disponivel);
-            System.out.println("------------------------------");
+            if (isValidObjectId(idDocumento)) {
+                String nome = documento.getString("nome");
+                String cpf = documento.getString("cpf");
+                String telefone = documento.getString("telefone");
+                String placaMoto = documento.getString("placaMoto");
+                String disponivel = documento.getString("disponivel");
+
+                System.out.println("Identificador: " + idDocumento);
+                System.out.println("Nome: " + nome);
+                System.out.println("CPF: " + cpf);
+                System.out.println("Telefone: " + telefone);
+                System.out.println("Placa da moto: " + placaMoto);
+                System.out.println("Disponível: " + disponivel);
+                System.out.println("------------------------------");
+            } else {
+                JOptionPane.showMessageDialog(null, "ID inválido: " + idDocumento);
+            }
         }
     }
+
+
 
     private static boolean motoboyExiste(String id) {
         Document motoboy = collection.find(getFiltro(id)).first();
@@ -129,6 +145,17 @@ public class MotoboyRepository {
     private static Bson getFiltro(String id){
         return Filters.eq("_id", new ObjectId(id));
     }
+    private static boolean isValidObjectId(String id) {
+        if (id == null || id.length() != 24) {
+            return false;
+        }
 
+        try {
+            new ObjectId(id);
+            return true;
+        } catch (IllegalArgumentException e) {
+            return false;
+        }
+    }
 
 }
